# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kitelepítési Támogató Rendszer API",
        "description": "QR-kódos kitelepítési, regisztrációs és befogadóhelyi nyomon követő prototípus REST API-ja.\n\nAz autentikáció Laravel Sanctum SPA session-cookie alapú: előbb GET /sanctum/csrf-cookie, utána POST /api/login. A böngészőn kívüli kliensek (pl. ez a Swagger UI) ezért a `/api/login` végpontot csak akkor tudják sikeresen meghívni, ha a kliens ugyanabból az origin-ből fut, ahonnan a cookie-kat is fogadja.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Helyi fejlesztői szerver"
        }
    ],
    "paths": {
...
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Szállítójármű-flotta törzsadatainak listája, aktuális hozzárendeléssel",
                "operationId": "043b3ce51f006f946511103cfe1c2ca1",
                "responses": {
                    "200": {
                        "description": "Járművek listája"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Új jármű felvétele a flottába",
                "description": "Csak admin és vezető szerepkör jogosult.",
                "operationId": "74e9ae3e5f1ee2f69f3a177820d204b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plate_number",
                                    "label"
                                ],
                                "properties": {
                                    "plate_number": {
                                        "type": "string",
                                        "example": "AAA-123"
                                    },
                                    "label": {
                                        "type": "string",
                                        "example": "1. sz. busz"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "driver_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Létrehozott jármű"
                    },
                    "403": {
                        "description": "Nincs jogosultság"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        },
        "/api/vehicles/{vehicle}": {
            "put": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Jármű törzsadatainak módosítása",
                "description": "Csak admin és vezető szerepkör jogosult.",
                "operationId": "82a1b2a717b16eb7e08a86288a7f8093",
                "parameters": [
                    {
                        "name": "vehicle",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Frissített jármű"
                    },
                    "403": {
                        "description": "Nincs jogosultság"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Jármű törlése a flottából",
                "description": "Csak admin és vezető jogosult, és csak akkor, ha a jármű jelenleg nincs folyamatban lévő eseményhez rendelve.",
                "operationId": "495952c6402028fab9f82c080af676ad",
                "parameters": [
                    {
                        "name": "vehicle",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sikeres törlés"
                    },
                    "403": {
                        "description": "Nincs jogosultság"
                    },
                    "409": {
                        "description": "A jármű jelenleg használatban van"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctumSession": {
                "type": "apiKey",
                "description": "Laravel Sanctum SPA session-cookie autentikáció. Bejelentkezés: POST /api/login.",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}